+++ example/lib/home_page.dart
@@ -1,7 +1,6 @@
-import 'dart:math';
-
 import 'package:flutter/material.dart';
 import 'package:flutterfly/flutterfly.dart';
+import 'package:flutterfly_example/global_state.dart';
 import 'details_page.dart';
 
 class HomePage extends StatefulWidget {
@@ -10,26 +9,6 @@ class HomePage extends StatefulWidget {
 }
 
 class _HomePageState extends State<HomePage> {
-  List<ThumbnailModel> _storeItems = [];
-
-  @override
-  void initState() {
-    super.initState();
-    for (int i = 1; i <= 7; i++) {
-      _storeItems.add(ThumbnailModel(
-          id: i,
-          name: 'The Greatest Lego of All Time $i',
-          imgUrl: 'https://randomuser.me/api/portraits/lego/$i.jpg',
-          price: Random().nextInt(1000)));
-    }
-    // add null img for testing
-    _storeItems.add(ThumbnailModel(
-        id: _storeItems.length + 1,
-        name: 'A Rejected Lego',
-        imgUrl: null,
-        price: Random().nextInt(1000)));
-  }
-
   @override
   Widget build(BuildContext context) {
     return CustomScrollView(primary: false, slivers: <Widget>[
@@ -48,11 +27,9 @@ class _HomePageState extends State<HomePage> {
 
   List<Widget> _buildStoreList(BuildContext context) {
     List<Widget> cards = [];
-    for (ThumbnailModel item in _storeItems) {
+    for (StoreModel item in GlobalState().storeModels) {
       cards.add(ThumbnailStore(
-        imgUrl: item.imgUrl,
-        title: item.name,
-        price: item.price,
+        model: item,
         onTap: () {
           Navigator.push(
             context,
@@ -65,28 +42,13 @@ class _HomePageState extends State<HomePage> {
   }
 }
 
-class ThumbnailModel {
-  final int id;
-  final String name;
-  final String imgUrl;
-  final int price;
-
-  ThumbnailModel({this.id, this.name, this.imgUrl, this.price});
-}
-
 class ThumbnailStore extends StatefulWidget {
-  final String imgUrl;
+  final StoreModel model;
   final double imgAspectRatio;
-  final String title;
-  final int price;
   final void Function() onTap;
 
   ThumbnailStore(
-      {@required this.imgUrl,
-      @required this.title,
-      @required this.price,
-      @required this.onTap,
-      this.imgAspectRatio = 1.0});
+      {@required this.model, @required this.onTap, this.imgAspectRatio = 1.0});
 
   @override
   _ThumbnailStoreState createState() => _ThumbnailStoreState();
@@ -108,14 +70,14 @@ class _ThumbnailStoreState extends State<ThumbnailStore> {
                 padding: EdgeInsets.all(10),
                 height: 60,
                 child: Text(
-                  widget.title,
+                  widget.model.name,
                   maxLines: 2,
                   overflow: TextOverflow.ellipsis,
                   textAlign: TextAlign.center,
                   style: TextStyle(fontSize: 16),
                 )),
             FlyImage(
-              url: widget.imgUrl,
+              url: widget.model.imgUrl,
               aspectRatio: widget.imgAspectRatio,
             ),
             Expanded(
@@ -123,7 +85,7 @@ class _ThumbnailStoreState extends State<ThumbnailStore> {
                     alignment: Alignment.center,
                     padding: EdgeInsets.all(5),
                     child: Text(
-                      '${widget.price} USD',
+                      '${widget.model.price} USD',
                       textAlign: TextAlign.justify,
                       style:
                           TextStyle(fontWeight: FontWeight.bold, fontSize: 14),