+++ lib/src/form/components/index.dart
@@ -1,3 +1,4 @@
 export 'checkbox.dart';
 export 'datepicker.dart';
 export 'dropdown.dart';
+export 'numeric.dart';